pyproject.toml
[project]
name = "transcribee"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "faster-whisper",
    "torch",
    "torchaudio",
    "requests",
]

[project.optional-dependencies]
openai = ["openai>=1"]
anthropic = ["anthropic>=0.25"]
resemblyzer = ["resemblyzer", "scikit-learn"]
pyannote = ["pyannote.audio"]
dev = ["pytest", "openai>=1", "anthropic>=0.25", "resemblyzer", "scikit-learn"]

[project.scripts]
transcribee = "transcribee.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
